Update translate_table.py
I was working off of a copy of this script I had in the group directory and did not realize someone had also changed the order of the if statement checks from the original. 

The correct way to do it is the way Nick originally had, so I've changed this if statement back to how it was. (nb lists are shorter, bonds angles and dihedrals are the same length)